# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Instagram API Explorer backend
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SessionStatus {
  active: boolean;
  username?: string;
  userId?: string;
  csrfToken?: string;
  hasSessionId?: boolean;
}

export interface SessionInput {
  sessionId: string;
  csrfToken: string;
  username?: string;
  userId?: string;
  dsUserId?: string;
}

export type ApiResponseData = { [key: string]: unknown };

export type ApiResponseRequestHeaders = {[key: string]: string};

export interface ApiResponse {
  success: boolean;
  data?: ApiResponseData;
  endpoint: string;
  method: string;
  statusCode: number;
  durationMs: number;
  requestHeaders?: ApiResponseRequestHeaders;
  timestamp: string;
  error?: string;
}

export type GraphQLInputVariables = { [key: string]: unknown };

export interface GraphQLInput {
  docId: string;
  variables: GraphQLInputVariables;
}

export interface HistoryEntry {
  id: string;
  endpoint: string;
  method: string;
  statusCode: number;
  durationMs: number;
  timestamp: string;
  success: boolean;
  label: string;
}

export interface HistoryListResponse {
  entries: HistoryEntry[];
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  success: boolean;
  error: string;
}

export type GetProfileParams = {
username: string;
};

export type GetHashtagParams = {
tag: string;
after?: string;
};

export type GetUserFeedParams = {
userId: string;
after?: string;
};

export type GetCommentsParams = {
shortcode: string;
after?: string;
};