# lib/api-zod/src/generated/types/apiResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Instagram API Explorer backend
 * OpenAPI spec version: 0.1.0
 */
import type { ApiResponseData } from './apiResponseData';
import type { ApiResponseRequestHeaders } from './apiResponseRequestHeaders';

export interface ApiResponse {
  success: boolean;
  data?: ApiResponseData;
  endpoint: string;
  method: string;
  statusCode: number;
  durationMs: number;
  requestHeaders?: ApiResponseRequestHeaders;
  timestamp: string;
  error?: string;
}