# lib/api-zod/src/generated/types/graphQLInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Instagram API Explorer backend
 * OpenAPI spec version: 0.1.0
 */
import type { GraphQLInputVariables } from './graphQLInputVariables';

export interface GraphQLInput {
  docId: string;
  variables: GraphQLInputVariables;
}