# lib/api-zod/src/generated/types/historyListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Instagram API Explorer backend
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntry } from './historyEntry';

export interface HistoryListResponse {
  entries: HistoryEntry[];
}