# lib/api-zod/src/generated/types/sessionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Instagram API Explorer backend
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInput {
  sessionId: string;
  csrfToken: string;
  username?: string;
  userId?: string;
  dsUserId?: string;
}