# lib/api-zod/src/generated/types/sessionStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Instagram API Explorer backend
 * OpenAPI spec version: 0.1.0
 */

export interface SessionStatus {
  active: boolean;
  username?: string;
  userId?: string;
  csrfToken?: string;
  hasSessionId?: boolean;
}